Make Dockerfile compatible to latest upstream layout
With https://github.com/mumble-voip/mumble/pull/5838 the file structure
of the upstream repo has changed slightly and most importantly the
location of the default server ini file has changed as well. Thus, we
now have to check multiple locations inside our Dockerfile in order to
make it work with the old and the new layout.